refactor: extract media/lights output cluster from state.py God-Object (#1271)
Move the external-output method cluster — media transport (stop_media,
set_volume_on_player, seek_forward, play_deferred_song), party-lights
control (configure_party_lights, disable_party_lights, _lights_set_phase,
_lights_flash) and adjust_volume — out of game/state.py into a new
MediaControlMixin (game/state_media.py), following the existing
LeaderboardMixin / TtsAnnouncerMixin extraction pattern.

Behavior-preserving: GameState inherits the mixin, so its public API and
every caller/test are unchanged. state.py: 2602 -> 2510 lines (-92).
